refatoring - config bigquery and time partitioned - added inheritance and raw_table_definition
